feat(permission): add Owner short-circuit for permission evaluation
- Skip UserAndAgent rules when caller.user_id == "owner"
- Owner still subject to AgentOnly rules and default policies
- Non-owner behavior unchanged

Source: issue #661
Type: feat
